requirement.txt
anyio==3.6.1
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
fastapi==0.84.0
Flask==2.2.2
fonttools==4.37.1
gunicorn==20.1.0
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
starlette==0.19.1
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
Werkzeug==2.2.2
zipp==3.8.1